Modify: 프로필 컴포넌트에서 계정 활성화 토글 버튼(ProfileActiveToggle) 분리, 로비 페이지(Lobby.jsx)에 게임 헤더 적용 절대 경로 환경 설정에서 @shared 별칭 삭제
