+++ src/pages/Game/Lobby.jsx
@@ -1,7 +1,19 @@
 import React from "react";
 
+import { ContentWrapper, WideContent, Main } from "@/styles/CommonStyle";
+import GameHeader from "@/components/Game/Shared/GameHeader";
+import Footer from "@/components/Web/Shared/Layout/Footer";
+
 const Lobby = () => {
-  return <h1>Lobby Page</h1>;
+  return (
+    <ContentWrapper row="center" col="center">
+      <WideContent dir="col">
+        <GameHeader />
+        <Main></Main>
+        <Footer />
+      </WideContent>
+    </ContentWrapper>
+  );
 };
 
 export default Lobby;
